+++ src/locale.ts
@@ -53,6 +53,16 @@ export interface BooleanLocale {
   isValue?: Message;
 }
 
+export interface LocaleObject {
+  mixed?: MixedLocale;
+  string?: StringLocale;
+  number?: NumberLocale;
+  date?: DateLocale;
+  boolean?: BooleanLocale;
+  object?: ObjectLocale;
+  array?: ArrayLocale;
+}
+
 export let mixed: Required<MixedLocale> = {
   default: '${path} is invalid',
   required: '${path} is a required field',
@@ -111,7 +121,7 @@ export let object: Required<ObjectLocale> = {
 export let array: Required<ArrayLocale> = {
   min: '${path} field must have at least ${min} items',
   max: '${path} field must have less than or equal to ${max} items',
-  length: '${path} must be have ${length} items',
+  length: '${path} must have ${length} items',
 };
 
 export default Object.assign(Object.create(null) as {}, {