+++ test/string.js
@@ -227,6 +227,26 @@ describe('String types', () => {
     ]);
   });
 
+  it('should check allowed values at the end',() => {
+    return Promise.all([
+      string()
+        .required('Required')
+        .notOneOf([ref('$someKey')])
+        .validate('',{context:{someKey:''}})
+        .should.be.rejected().then(err => {
+          err.type.should.equal('required')
+        }),
+      object({
+        email:string().required('Email Required'),
+        password:string().required('Password Required').notOneOf([ref('email')]),
+      }).validate({email:'',password:''},{abortEarly:false})
+      .should.be.rejected().then(err => {
+        err.errors.should.include('Email Required');
+        err.errors.should.include('Password Required');
+      })
+    ]);
+  });
+
   it('should validate transforms', function () {
     return Promise.all([
       string().trim().isValid(' 3  ').should.eventually().equal(true),